refactor: update path to Summary component
